DJ/requirements/local.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
cfgv==3.3.1
distlib==0.3.6
Django==4.1.5
django-ckeditor==6.5.1
django-js-asset==2.0.0
filelock==3.9.0
identify==2.5.17
nodeenv==1.7.0
Pillow==9.4.0
platformdirs==2.6.2
pre-commit==3.0.4
PyYAML==6.0
sqlparse==0.4.3
virtualenv==20.17.1
